refactor: optimize conversation notification structure.
